requirements.txt
pyserial
scipy